pyproject.toml
[project]
name = "llm-tracing-demo"
version = "0.1.0"
description = "FastAPI LLM tracing demo with OpenRouter and Langfuse"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn",
    "openai",
    "langfuse",
    "python-dotenv",
    "pydantic",
    "pydantic-settings",
]
